info: refactor info_update()
Return early if we can so we don't have nested if statements. Use the
power of str.join() and list comprehensions. All callers of info_update
in main.py were never using the blank_window and skip_lyrics, so just
remove those arguments. Also the bitrate was in a try catch for
keyerrors, so just replace that to use dict.get() and check the status
for non None.